backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
supabase==2.3.0
redis==5.0.1
openai==1.3.5
paramiko==3.3.1
pydantic==2.5.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
aiofiles==23.2.1
python-dotenv==1.0.0